% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudySPoTApi.r
\name{StudySPoTApi_add_atomic_study}
\alias{StudySPoTApi_add_atomic_study}
\title{Create a single new object from a given data file with optional supplied metadata.}
\usage{
StudySPoTApi_add_atomic_study(source, ...)
}
\description{
Create a single new object from a given data file with optional supplied metadata.
}
