% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudySPoTApi.r
\name{StudySPoTApi_delete_study}
\alias{StudySPoTApi_delete_study}
\title{Delete a given study object. This will not delete related objects or links to related objects.}
\usage{
StudySPoTApi_delete_study(id, ...)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}
}
\description{
Delete a given study object. This will not delete related objects or links to related objects.
}
