% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudySPoTApi.r
\name{StudySPoTApi_get_study_versions}
\alias{StudySPoTApi_get_study_versions}
\title{Retrieve a list of object versions by ID}
\usage{
StudySPoTApi_get_study_versions(id, ...)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}
}
\description{
Retrieve a list of object versions by ID
}
