% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudySPoTApi.r
\name{StudySPoTApi_update_study}
\alias{StudySPoTApi_update_study}
\title{Update a study object}
\usage{
StudySPoTApi_update_study(id, body, ...)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}

\item{body}{Metadata in the form of \code{{key: value, key2: value2, ...}}}
}
\description{
Update a study object
}
